refactor: OrderService 의존성 리팩토링
